Move the account list page somewhere more sane
